requirements.txt
pylint==2.10.2
black==21.7b0
Pillow==9.0
xlwings==0.26.3
pykakasi==2.2.1